Add nitpick to the fix aggregators list
Nitpick is a tool that takes style definitions for your project configuration files, such as `.editorconfig`, `.pre-commit`, `.eslint.json` and keeps them all up to date.

Disclaimer: I have contributed to the nitpick project as we use it to maintain coding standards tooling across a tech company's repositories. I am not a maintainer of the project.